requirements.txt
django==3.2
django-filter==2.4.0
django-rest-swagger==2.2.0
django-restframework==0.0.1
django-role-permissions==3.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.5.0
Pillow==8.3.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
PyJWT==1.7.1

